final release for multithreading

+++ Headers/PrimeChecker.h
@@ -5,14 +5,14 @@
 class PrimeCheck
 {
 public:
-	long long int numberToCheck;
-	long long int check = 2;
-	long long int check2;
+	uint64_t numberToCheck;
+	uint64_t check = 2;
+	uint64_t check2;
 
-	void checkFirstHalf(long long int number);
-	void checkSecondHalf(long long int number);
-	void checkFull(long long int number);
-	bool checkPrimeTheEasyWay(long long int number);
+	void checkFirstHalf(uint64_t number);
+	void checkSecondHalf(uint64_t number);
+	void checkFull(uint64_t number);
+	bool checkPrimeTheEasyWay(uint64_t number);
 
 	std::vector<bool> primebits;
 private:
